requirements.txt
sqlalchemy~=2.0.31
fastapi~=0.111.0
asyncio
python-dotenv~=1.0.1
asyncpg
alembic~=1.13.2
psycopg2
uvicorn~=0.30.1
pydantic~=2.8.0
starlette~=0.37.2
python-slugify~=8.0.4
jose~=1.0.0
passlib~=1.7.4